+++ addon/index.js
@@ -1,0 +1,5 @@
+import IvyRedactorComponent from './components/ivy-redactor';
+
+export {
+  IvyRedactorComponent
+};